requirements.txt
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
coverage==7.2.3
exceptiongroup==1.1.1
faiss-cpu==1.7.3
filelock==3.12.0
flake8==6.0.0
huggingface-hub==0.13.4
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
lit==16.0.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy==1.2.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.15.1
pytest==7.3.1
pytest-cov==4.0.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rich==13.3.5
sympy==1.11.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
torchdata==0.6.0
torchtext==0.15.1
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
